LPS-78486 Fix packages
