+++ modules/apps/headless-apio/forms/forms-apio-api/build.gradle
@@ -2,6 +2,5 @@ sourceCompatibility = "1.8"
 targetCompatibility = "1.8"
 
 dependencies {
-	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
 	compileOnly project(":apps:foundation:apio-architect:apio-architect-api")
 }
