+++ modules/apps/headless-apio/forms/forms-apio-impl/bnd.bnd
@@ -1,5 +1,3 @@
 Bundle-Name: Liferay Forms Apio Implementation
 Bundle-SymbolicName: com.liferay.forms.apio.impl
 Bundle-Version: 1.0.0
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Site Management
