+++ modules/apps/headless-apio/forms/forms-apio-impl/src/main/java/com/liferay/forms/apio/internal/provider/FormInstanceRecordServiceContextProvider.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.forms.apio.internal.architect.provider;
+package com.liferay.forms.apio.internal.provider;
 
 import com.liferay.apio.architect.provider.Provider;
 import com.liferay.dynamic.data.mapping.model.DDMFormInstanceRecord;
-import com.liferay.forms.apio.internal.architect.FormInstanceRecordServiceContext;
+import com.liferay.forms.apio.internal.FormInstanceRecordServiceContext;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextFactory;
